Added a Program.cs and moved some things there from Challenge.cs Changed Challenge to use a function as a parameter for the data source, they could create the events in code, use a csv, db, w/e they want. Fixed some spelling/grammar Some suggestions for changes
(cherry picked from commit 23d73e8f64f3c9d8110f825862afb113f18f8b97)
