+++ README.md
@@ -1,7 +1,7 @@
 # Challenge
 Code repository for challenge boilerplate code.
 ### How to use this repository
-Fork the repository, clone into it and implement your challenge submission in the specific method.
+Fork the repository, clone into it and implement your challenge submission in the specified methods.
 
 ## Paid time challenge
-The boiler includes some helper classes for time sensitive data, all the surrounding test calls required and the method signature with empty implementation.
+The boiler includes some helper classes for time sensitive data, all the surrounding test calls required, and the method signature with empty implementation.
